pyproject.toml
[tool.poetry]
name = "SpotifyProject"
version = "0.1.0"
description = ""
authors = ["Jack McCain <jackmccain56@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
requests = "^2.25.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
